ci: Remove schedule (cron) trigger from build-buildboxes (#3879)
Remove the schedule (cron) trigger from the build-buildboxes workflow as
it does not work as I thought it did - I was expecting the workflow to
be called on each branch that the workflow exists on the given schedule.
However, scheduled workflows only run on the default branch (master).

Instead a follow-up commit will introduce the schedule trigger in a new
workflow that calls the build-buildboxes workflow on each of the
branches it needs to run on. The build-buildboxes is now a simple
callable and dispatchable workflow that just builds the buildboxes on
demand.

Remove the `amd64` exclusion logic as it seems a daily update of the
build boxes is sufficient and it does not need to be done on each push.
This means we do not need a trigger in the oss teleport repo for this
and we will no longer build the amd64 buildboxes there. It will all be
done from teleport.e in the build-buildboxes workflow, so we don't need
the arm64-only logic.

Source-commit: gravitational/teleport.e@4adf051ef8c80b0b9a395f67d9de2a4e0010e429
Source-pull-requests: gravitational/teleport.e#3879