Add node:rw to RoleDiscovery to allow EICE
This PR adds a missing permission to the RoleDiscovery so that it can
create nodes of EICE kind.
These nodes are the only ones created by the DiscoveryService.
Teleport Nodes are created when the teleport process joins the cluster
Agentless Nodes are either created via `tclt` or `teleport join` command