Convert WithRootClusterClient to use ClusterClient (#40168)
Replaces the deprecated SSH connection mechanism to the cluster
in favor of using the gRPC mechanism. The WithRootClusterClient
has also been deprecated as it requires at worst two dials to
a Teleport cluster for a single function. Callers should instead
explicitly create and reuse clients where possible so that no
extraneous dialing is required for each action being performed
against a cluster.