Prevent moderated session tests from blocking on read forever
Exposes SetReadDeadline on the terminal web socket so that the
test helper `waitForOutput` can set it to unblock reads if the
timeout is exceeded. While this alone should not fix any
flaky tests it should allow tests failures to occur prior to
the 10m test timeout and isolate at which parts of the test things
are becoming stuck.

Updates #36061 and #30892