Fix EAS Token generation (#40154)
In the latest release, we added support for S3 bucket as OIDC Provider
URL.
There are two places were the Token is generated.
One used in Discover Wizard, EICE access and Auto-Discover.
Another one for the External Audit Storage.

The first was updated to include the S3 Bucket as possible Issuer.
EAS flow was not updated.

This PR merges both flows.