Refactors rust lint out into its own job which runs on a bare GH runner (#40146)
* Refactors rust lint out into its own job which runs on a bare GH runner

Also adds -s to silence all make commands in workflows which are used as
some form of output in the workflow script. This is to prevent the
workflow from failing due to noise from the make command (such as when
make tells you it's entering a directory, etc).

Also updates to latest e-ref

* Removes `--no-print-directory` where `make` output is used as a variable
replaceing it with `-s` instead. `-s` is a universal POSIX standard and
makes sense in all such cases.

https://www.gnu.org/software/automake/manual/html_node/Tricks-For-Silencing-Make.html

* aligning e with master

* aligning e with master

* aligning e with master