[v13] build: Cache buildbox layers inline (#40241)
Use the `docker buildx` `--cache-to type=inline` to put the build cache
in the buildbox image, to with with the existing `--cache-from`
parameter. To do this we need to use `docker buildx`, and to ensure that
a buildbox image is available locally if a container builder is being
used, ensure `--load` is used if we are not pushing to a remote
registry, based on the `PUSH` make variable.

This rolls up a number of changes made on master that are intertwined
with other changes we do not want to backport. This smaller set of
changes aligns the way the buildboxes are built, cached and pushed
across the branches without bringing across the larger scale changes for
PIC/PIE builds and ARM64 FIPS builds.

This also allows us to remove the `PULL_ON_CI` make "macro" since docker
will effectively do the same with `--cache-from` when those images have
an inline cache.