[v15] build: Update darwin-signing for new GitHub Actions environment names
Update `darwin-signing.mk` to accept `build-prod` and `build-stage` as
environment names, leaving `promote` and `build` as deprecated, soon to
be removed. For this, `CLEAN_ENV_NAME` is introduced and signing
variables are defined in terms of that, where hyphens in the
`ENVIRONMENT_NAME` are replaced by underscores. This keeps the make
variables names as valid environment variable names.

Add a `print-darwin-signing-vars` target to print out the signing
variables in use for debugging purposes.