ci: Add scheduled rebuild of all buildboxes (#3892)
Add a GitHub Actions scheduled job to rebuild the buildboxes of all
active branches to be run nightly. Run them serially so we don't need a
lot of self-hosted arm64 runners at the same time.

Also include a disabled job for when the next release gets cut but
before it is actually released.

Source-commit: gravitational/teleport.e@d57058b71b564de25acd6c291652f9d0c9dfb1b1
Source-pull-requests: gravitational/teleport.e#3892