player: fix seeking during dead time (#38914)
* player: fix seeking during dead time

This addresses a bug in the new player API where the player does
not respond to seeking during large periods of "dead time."
For example, if there is 30 seconds of dead time between two events,
the player is effectively in a 30 second sleep. Attempting to change
the position of playback during this 30-second period works, but is
not observed until the sleep completes.

With this change, the sleep period is canceled when the user seeks
to a new position.

Fixes #38560

* Improve UX when the user seeks while paused

Restart the stream in this case rather than trying to keep server
and client state in sync.

* Fix

* Update desktop player

* Lint/test fixes

* increase timeout

* Use a real clock for the new test