Connections: Avoid creating component during rendering & remove non-reusable hook (#40347)
* Extract Connections slider step to separate component

This avoids creating a fresh component on each render of Connections.
Otherwise there are weird issues with focus and component identity.

* Remove isAnyConnectionActive and activateItem from useConnections

* Use useConnections only in ConnectionsSliderStep

useConnections calculates some stuff each time it's called, so it doesn't
make sense to call it from two different components. The next commit is
going to get rid of this hook anyway.

* Add tests for preserving connection position

* Get rid of useConnections hook

* Remove "on" prefix from func props that do very specific things