Fix UI workflows (#3948)
* Fix UI workflows

We need to have both the OSS and enterprise repos in order for
the changes check to function correctly.

* ci: Add read permission for dorny/paths-filter checkout

Add the `contents: read` permission to the job for `dorny/paths-filter`
so that we can checkout teleport.e. This is needed as we lose that
permission if any other is specified (`pull-requests: read`), and
teleport.e is not public.

Also change the `if` condition to only not checkout if the event is a
pull request as only for that case does the paths filter access the
changed files via the GitHub API. All other cases require a checkout.

---------

Co-authored-by: Cam Hutchison <camh@goteleport.com>

Source-commit: gravitational/teleport.e@e966ca76eb657e8fbc9dbcd40cf6aa1c36222a4a
Source-pull-requests: gravitational/teleport.e#3948