Launch VNet when connecting to a TCP app
Lots of going on in this commit.

connectToApp used to take launchInBrowserIfWebApp. This is because
connectToApp was used by both the search bar and the unified resources
to handle connecting to an app.

In the unified resources, it set up a gateway since opening the app
in the browser was handled through clicking a link. In the search bar,
it launched the app in the browser since that's what we want to happen
when selecting a web app through the search bar.

Now connectToApp always launches web apps in the browser and creating
app gateways has been moved to setUpAppGateway in the same file.