ci: Build other branch buildboxes even if others fail (#3935)
Add an `if: always()` to the buildbox jobs so we continue to build the
other branches even if earlier one(s) fail. Failures are typically for
transient reasons, so there's no need to stop the lot if some fail.
Other reasons are likely to be branch-specific so again, no need to
stop. The dependencies exist only for sequencing so we don't do too many
concurrent builds on our self-hosted runners - they are not build
dependencies.

Source-commit: gravitational/teleport.e@6bd8b6eb1d5e1a804fd89b9d6698753b1450811f
Source-pull-requests: gravitational/teleport.e#3935