[v14] build: Simplify darwin-signing and add new GHA environments (#40388)
* [v14] build: Sign mac artifacts using key name not hash

Tidy up the mac signing makefile magic to sign using the key name and
not the key hash. Since we look up the key has by the name, there's no
benefit to using the hash as all the signing commands will take the key
name as well. This removes a bunch of fragility around exporting
variables, make errors and running the key lookup commands only on a mac
that it leaves it quite a bit simpler.

The `CSC_NAME` variable used by electron-builder to build Teleport
Connect does not want the full key name - it wants the purpose part
stripped off the front (`Developer ID Application:`), so we do that. The
other signing commands take the full key name.

This also simplifies the `NOTARIZE_BINARIES` set of commands, allowing
them to be broken up removing the `&&` conjunction. This makes it clear
exactly what command fails if there is a failure as make runs them
one-by-one instead of all at once. It also allows the Makefile in
teleport.e a little simpler as it no longer needs to run the
notarization from the parent directory.

* [v14] build: Update darwin-signing for new GitHub Actions environment names

Update `darwin-signing.mk` to accept `build-prod` and `build-stage` as
environment names, leaving `promote` and `build` as deprecated, soon to
be removed. For this, `CLEAN_ENV_NAME` is introduced and signing
variables are defined in terms of that, where hyphens in the
`ENVIRONMENT_NAME` are replaced by underscores. This keeps the make
variables names as valid environment variable names.

Add a `print-darwin-signing-vars` target to print out the signing
variables in use for debugging purposes.

* Update e ref for mac signing updates