Add a linter for `[Resource] Access` (#40421)
This is a legacy convention for describing Teleport features. This
change ports the linter for `[Resource] Access` usage from
`gravitational/docs` to a vale style (we'll need to remove the linter
from `gravitational/docs` separately). It also excludes "Database Access
Controls" from the linter, since Database Access Controls is a current
Teleport feature.