[v14] Update rust to latest stable (1.77.0) (#39992)
* Updates `RUST_VERSION` to the latest stable (1.77.0)

* Fixes linting errors

* Comment out #[allow(clippy::arc_with_non_send_sync)]

This is a temporary hack to get CI to pass. It is necessary for the
RUST_VERSION which we are updating to (1.77.0), however CI is still
using an image with the previous version 1.71.1. Meanwhile `arc_with_non_send_sync`
was added in 1.72.0 (https://rust-lang.github.io/rust-clippy/master/index.html#/arc_with_non_send_sync).

The PR with this commit will therefore require a follow up PR to uncomment
the line once the CI image is updated.