Fix context usage in DiscoveryService initialization
The created context is the one that should be used for context
propagation.

Most of this doesn't change anything, except for the AccessGraph, where
we span a goroutine using the received context.
It was using the initial one (`process.ExitContext`), but now it will
use the wrapped context which respects the `s.cancelFn`.