ci: Build the centos7-assets buildbox weekly (#3936)
Add an additional buildbox to the `build-buildboxes` workflow with some
conditional logic to build the centos7-assets buildbox once per week on
Sundays. Currently this is done on master only in the OSS teleport
repository. We need to do it here because we need to build AMD64 and
ARM64 versions and the latter can only be done on our self-hosted
runners only available from teleport.e.

Turn off `fail-fast` as one buildbox failure is unlikely to mean the
others will fail given each builds with a different dockerfile.

Source-commit: gravitational/teleport.e@a9e1c3629d2241da9a665e1e788a1c6d60125474
Source-pull-requests: gravitational/teleport.e#3936