[v15] fix: return `OpaqueAccessDenied` for `NotFound` remote cluster errors (#40681)
Backport #40571 to branch/v15

This commit modifies OpaqueAccessDenied to return an identical generic
NotFound error whether the input error is NotFound or AccessDenied.
The commit also updates all call sites of OpaqueAccessDenied to use it in
the paths where there is any error fetching the resource as well as when
access is denied.

It doesn't do much good to return AccessDenied errors as NotFound, if
they don't match the NotFound error you would get if the resource really
didn't exist.
It's trivial to tell the errors apart and discover the existence of a
resource you shouldn't be allowed to access.
The commit attempts to mitigate that issue and properly hide the
existence of resources the user should not be allowed to list.

Changelog: generic "not found" errors are returned whether a remote
cluster can't be found or access is denied.

---------

Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>