SCIM Group handling (#3529)
* Okta User & AccessList sync changes to implement SCIM groups

This PR makes some changes to the Okta AccessList and User sync
services to share reconcoler logic with the SCUM service. This
work was originally part of the SCIM group support PR, but was
split put for easier review.

Changes include:
 - replacing `okta.IsOktaUserInOrg()` with the generic `MatchByLabels()`,
   and dealing the knock-on effects of that change in Okta and SCIM
 - Cutting down the AccessList interface used by the AccessListMember
   reconciler in order to make it easier to mock in the SCIM package
 - Replacing the map+mutex pairs and manual (un)locking in the
   `accessListSync` type with `utils.SyncMap`

Changelog: None

* Groups SCIM support

Implments support for SCIM groups, which are translated to Teleport
AccessLists and managed as such. Currently only Okta groups are
supported, and the AccessLists it generates are compatible with
those cretaed by the Okta Sync Service.

Changelog: Adds SCIM support for Okta

* Review tidy

* Moved resourceToAcceesList from shim to main body

* remove debug output

* Review advice

* Allow IdP to customize AccessListMember records

 - Allows IDP to suggest role names
 - Allows IDP to set AccessListMember records

* Linter appeasement

* Update groups.go

* Fixes member deletion in Okta-derived access list (#3787)

The SCIM service was creating AccessList definitions with some empty
fields set to `nil`, where the Okta sync service wuld use an empty
structure instead.

This difference ran (potentially?) spuriously triggered the the
change validation code for Okta-derived `AccessList`s in the access list
service.

This patch updates the SCIM service to create `AccessList` records in the
Sync Service style, which no longer trips the change validator and
allows members to be deleted from an Okta-derived AccessList

* Simplifies SCIM group reconciliation

 - Uses the reconciler build in to the AccessList service rater than
   creating yet another reconciler. This is a win for both code size,
   and elimiating a race condition where an AccessList could be modified
   by another process during reconciliation, resulting in an
   inconsistent AccessList state
 - Simplifies tests to match

* Locking and preserve member values in SCIM push

* Fix typo

* Revert access list interval

* Preserve AddedBy, use cached Users service

* Properly plub through member options

* Make preserving access list metadata the default

* refactor shim

* refactor shim

Source-commit: gravitational/teleport.e@7821ee6256a87242f7deea1ad4530122e8695c19
Source-pull-requests: gravitational/teleport.e#3529