Makefile cleanup (#40769)
* Remove goinstall make target

This target hasn't been updated in 7+ years and installs a version
of Teleport that won't actually run due to missing build tags.

* Use 'go run' to invoke addlicense

The go tool is smart enough not to download and build the tool if
the correct version is available, and this eliminates the need for
us to ask what the GOPATH is.

* Misc Makefile cleanup

Several new targets were added in the middle of sections that should
not have been broken up.

* Remove outdated comment

make all _does_ include webassets

* Remove old reference to GIT variable