Extend target resolution for ssh connections to use predicates and fuzzy search
tsh had the ability to connect to a host by name, ip, or by matching
on a set of provided labels. However, if the target needed to be
identified by some other data associated with the types.Server it
required a two step process involving tsh ls:

`tsh ssh foo@$(tsh ls --search=<needle> --format json | jq -r \'.[0].metadata.name\')`

Not only is this a lot to type and remember, but it also requires
that jq be present to work. To make this simpler on end users, tsh
now supports --search and --query directly. The above can be invoked
directly via:

`tsh ssh --search=<needle>` or `tsh ssh --query='labels.foo == "bar"`

In addition to updating tsh ssh, this also adds the same changes to
tsh proxy ssh. The only difference between the two is the behavior
for when there are multiple matches. If tsh ssh --search/--query
has multiple matches the behavior is the same as today when
tsh ssh user@foo=bar has multiple matches. tsh proxy ssh however is
limited to a single match, if more than one server are found then
and error is returned and the process is aborted.

To unlock the full power of tsh proxy ssh --query/--search, the tsh
proxy_templates have also been extended to include a corresponding
Search and Query field. The order of precedence for resolving hosts
in proxy_templates is Host, Search, Query.