Remove Assist command execution and chat feature (#40806)
* Remove Assist command execution and chat feature

The changes remove tests and related code for command execution in the assist package and web command handling.

* Remove Teleport Assist chat from CHANGELOG.md

The changes in this commit reflect the removal of the Teleport Assist chat from the CHANGELOG documentation of Teleport version 16. The Assist feature continues to be available in the SSH Web Terminal and Audit Monitoring.

* Remove assist chat from the test plan