Enroll Mode for EC2 AutoDiscovery types
We have two enroll modes for EC2 instances:
- script
- eice

Currently, if an Integration exists we always use the EICE mode.
However, we might want to use the Integration credentials to run the SSM
commands and install teleport on the target instance.

This PR adds a new knob in the DiscoveryConfig that allows us to
configure the enrollment mode when an integration is configured.