Implement autostart for VNet in Connect (#40900)
* Refactor tests that depended on workspace service state

Those tests explicitly set the entire WorkspacesService state and would
have to be modified each time we added something to the state.

There's no reason for them to know about WorkspacesService state, so this
commit refactors them out of that knowledge.

* Add useAppState

* Automatically start VNet

* subscribeWithSelector: Return unsubscribe function

The unsubscribe function will be needed in order to comply with
useSyncExternalStore API in the next commit.

* Add isInitialized to WorkspacesService

* Add useImmutableStore

* Wait for workspace init before VNet autostart

* Rename useAppState to usePersistedState

* Rename useImmutableStore to useStoreSelector