S3: abort uploads which are already present in the bucket
This change addresses some 'upload not found' errors we've observed
when using S3 as a session recording backend in high-availability
clusters. Since these clusters run multiple auth servers, and there
is no coordination between the upload completers in each auth server
(at least until #40926 merges), it's possible for multiple separate
uploads to exist for the same session recording. In this case, the
S3 API returns a "not found" error even though the recording file
is already present in S3, which causes the upload completer to
continuously try to complete the upload.

Fix this by checking if the uploaded file is already present.
If it is, we can safely abandon the current upload, which will
prevent the endless retry loop that we see today.

Note: for now, the abort functionally is opt-in and requires
the TELEPORT_ABORT_COMPLETED_UPLOADS environment variable to be
set to "yes". If unset, Teleport will log the uploads it would
have aborted, but will otherwise do nothing.

Additionally, we also stop manually decrementing the incomplete
uploads prometheus metric. There are typically multiple auth servers
running, so attempting to guess at how many uploads remain whithout
knowledge of what the other auth server is doing is prone to error.