Extract a shared part of access requests (part 1) (#4041)
* Move all access requests files from `Workflow` to `AccessRequests`

* Extract a shared part of access requests

* Move `ResourceKind` and `ResourceMap` to the shared `AccessRequests`

* Move `RequestList` component back to `Workflow`

I didn't notice that Connect and Web UI have their own resource lists.
They only use some common cells that I kept in the shared `AccessRequests`.

* Remove unnecessary `AccessRequestWithFlags` from the shared `AccessRequests` and teleterm

It also turned out that teleterm's `RequestList`
was implicitly expecting the requests to be of type `AccessRequestWithFlags`
(`request as Row`, so tsc didn't see any problems when I changed the type in tests).

I refactored it, so now the flags are passed separately, and the type assertion is gone.

* Pass `requestId` from props to `RequestView`

* Fix `ResourceKind` deprecation comment

* Extract a shared part of access requests (part 2) (#4042)

* Move the common `AccessRequest` (and subtypes) from `services/workflow` to `services/accessRequests`

And the same for `makeAccessRequests`.

* Remove imports from access lists

Probably it would be better to move the referenced stuff too,
but the PRs are already large.
Additionally, we only import two things, I think it's not that bad to inline them.

* Remove remaining `e-teleport` imports from `AccessRequests/Shared`

* Rename stories' titles

Source-commit: gravitational/teleport.e@9e97d275425238cb77f4f0a9d0254a5cb042c86d
Source-pull-requests: gravitational/teleport.e#4041