Add a retry mechanism to `UploadCompleter`'s `session.end` checker
When uploading session assets, `UploadCompleter` checks if the stale
upload contains the `session.end` event. This check is done using
`StreamSessionEvents` call to Auth Server to retrieve all the events
recorded for a particular session. When the session misses the
`session.end` event, the upload completer fakes it and creates it in
order to Teleport being able to display the session recordings. Teleport
requires `session.end` to be able to display them correctly.

The upload completer after finalyzing the upload, schedules a goroutine
to run 2min after to properly range all events in the session and
creating the event if missed.

The issue happens if the upload completer can successfuly complete the
upload - auth is operational and s3 storage is also operationa -l but
audit events backend isn't available at the time. In those cases,
although the recording was correctly completed, emiting a `session.end`
event will fail and no-one will be able to retrieve or analyze the
session recordings content.

This PR adds a logic to keep ensuring the session properly has an end
event until the audit log backend is operational again.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>