auth: require a semaphore in order to complete uploads (#41103)
Having multiple auth servers attempt to complete uploads against the
same shared backend results in racy behavior. This code is focused
on cleanup and is not performance sensitive, so we leverage a
semaphore to make sure only one auth server is attempting to complete
uploads at any point in time.

This new behavior can be disabled with an environment variable,
which will provide an escape hatch in case we we encounter
problems with it.