Add a topic page generator and proof of concept
Add a script that generates a partial that lists all pages within a
directory tree. We can include this partial in an "All Topics" page to
reduce complexity in the navigation sidebar. Rather than list all guides
in the sidebar, we can focus on high-priority ones while directing users
looking for additional topics to the "All Topics" page.

As we reduce the number of navigation sidebar sections, we can begin to
lean on All Topics pages more heavily so we can feature high-priority
pages in the sidebar.

As a proof of concept, this change removes navigation sidebar links from
the "Manage your Cluster" section, and simplifies the introduction page
for this section, and adds a link to an "All Topics" page for "Manage
your Cluster".

This change also moves the Diagnostics configuration guide, which allows
the "Manage your Cluster" section to abide by the expectations of the
topic page generation script.

The change also renames the "Dynamic Resources" section to
"Infrastructure as Code" to be more meaningful to readers.