web: ensure active sessions show up correctly (#41221)
This ensures that users who have permissions to join active sessions
see the active sessions option in the web UI, even if they have a
wildcard deny rule.

Note: this does not make any changes to the RBAC system, it only
brings the web UI logic in sync with the existing RBAC system.

Closes gravitational/security-findings#26