Add Entra ID directory reconciler (#4008)
* Add Entra ID directory reconciler

This adds a reconciler for synchronizing the following:
* Entra ID users as Teleport users
* Entra ID groups as Teleport access lists, including user memberships

This does not currently support nested Entra groups,
as support for nested access lists in Teleport is not yet in `master`.

This does not include syncing objects only relevant to Access Graph,
such as SSO connectors, or the connection to Access Graph itself.

The reconciler is not wired up to anything yet
(the Entra ID plugin boilerplate PR is in review).

It is covered by unit tests, and has also been verified
by running it manually against a local instance.

* Avoid nil error with unconvertable members

* Grant the correct trait according to Entra group

* Set member join time to the current time

`preserveAccessListMemberMetadata` will make sure it is not updated
between reconciles

* Address nits

* Pass context to slog where available

* Do not preserve revision for access lists

Upsert() takes care of it

* Use proper page size when paginating

* Use ListUsers

* Use ListAccessLists

* Improve documentation for slugify, rename

* Improve nil handling for Entra ID directory objects

* Refactor if-else into a type switch

* Use an early-return

* Elaborate on resourceNameClean

* Wire up Entra directory reconciler to plugin

* Address lints

* Address nits

* Apply suggestions from code review

Co-authored-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

* Make internally-used interface types private

* Add godocs to public types

* Remove usage of the exp/maps package

* Avoid unnecessary Sprintf

---------

Co-authored-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

Source-commit: gravitational/teleport.e@7711c33a41c57c6b5c4826933d28deb17c4c1868
Source-pull-requests: gravitational/teleport.e#4008