[v15] Enroll Mode for EC2 AutoDiscover (#41365)
* Enroll Mode for EC2 AutoDiscovery types (#40937)

We have two enroll modes for EC2 instances:
- script
- eice

Currently, if an Integration exists we always use the EICE mode.
However, we might want to use the Integration credentials to run the SSM
commands and install teleport on the target instance.

This PR adds a new knob in the DiscoveryConfig that allows us to
configure the enrollment mode when an integration is configured.

* EC2 AutoDiscover Script mode with Integration (#40959)

This PR adds auto-discover support for EC2 installation when using the
Script enroll mode with Integration credentials.

This mode uses SSM commands to install teleport on the target host.
Previsouly, integration credentials would only allow the EICE mode.
From now on, we can use integration credentials to run the SSM commands
on the target node.