Fetch resources using `ListUnifiedResources` in the search bar (#41357)
* Get rid of `clusters.App` type

We shouldn't have added it, `addrWithProtocol` is calculated from other `tsh.App` properties, so we can just do it when needed.
This type prevents assigning `listUnifiedResources` response to the type returned from `searchResources`.
Also, this change will make setting up mocks and tests simpler.

* Remove resource kind from `ResourceSearchError`

We will fetch all resources for a cluster in a single request, so it is no longer needed.

* Replace separate calls for resources with `listUnifiedResources`

* In preview mode, fetch 20 items instead of 5

Previously we fetched 5 items x 4 resource kinds.

* Move `getAppAddrWithProtocol` to `services/tshd/app.ts`

* Extract `MAX_RANKED_RESULTS`